zl_test_server/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.10.1
idna==2.8
mysqlclient==1.4.2
pycryptodomex==3.7.2
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
python-memcached==1.59
pytz==2019.1
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.7
